+++ zephyr/lib/libc/minimal/source/string/string.c
@@ -229,11 +229,11 @@ void *memcpy(void *_Restrict d, const void *_Restrict s, size_t n)
 	unsigned char *d_byte = (unsigned char *)d;
 	const unsigned char *s_byte = (const unsigned char *)s;
 
-	if ((((unsigned int)d ^ (unsigned int)s_byte) & 0x3) == 0) {
+	if ((((unsigned long)d ^ (unsigned long)s_byte) & 0x3) == 0) {
 
 		/* do byte-sized copying until word-aligned or finished */
 
-		while (((unsigned int)d_byte) & 0x3) {
+		while (((unsigned long)d_byte) & 0x3) {
 			if (n == 0) {
 				return d;
 			}
@@ -279,7 +279,7 @@ void *memset(void *buf, int c, size_t n)
 	unsigned char *d_byte = (unsigned char *)buf;
 	unsigned char c_byte = (unsigned char)c;
 
-	while (((unsigned int)d_byte) & 0x3) {
+	while (((unsigned long)d_byte) & 0x3) {
 		if (n == 0) {
 			return buf;
 		}
